TESTS: consolidate polars OneHotEncoder tests into test_preprocessing
Merge test_fit_transform across backends via a dataframe_backend fixture
that folds the pandas infer_string option contexts into the backend axis,
so pandas keeps running once per option context and polars once. Tests
whose pandas and polars variants assert different things are moved
verbatim instead of merged. Collected test count (43) and coverage are
unchanged.
